code/MessageGenerator/requirements.txt
attrs==22.2.0
beautifulsoup4==4.12.2
branca==0.6.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.6
et-xmlfile==1.1.0
Fiona==1.9.1
folium==0.14.0
geographiclib==2.0
geojson==3.0.1
geopandas==0.12.2
geopy==2.3.0
google==3.0.0
greenlet==2.0.2
grpcio==1.51.3
grpcio-tools==1.51.3
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.2
munch==2.5.0
numpy==1.24.2
openpyxl==3.1.0
packaging==23.0
paho-mqtt==1.6.1
pandas==1.5.3
protobuf==4.22.1
psycopg2==2.9.5
pyproj==3.4.1
python-dateutil==2.8.2
pytz==2022.7.1
requests==2.28.2
shapely==2.0.1
six==1.16.0
soupsieve==2.4.1
SQLAlchemy==2.0.2
typing_extensions==4.4.0
urllib3==1.26.14
websockets==10.4
